pyproject.toml
[tool.poetry]
name = "digimon-card-game"
version = "0.1.0"
description = ""
authors = ["ezeBalsamo <alexander.ezequiel.balsamo@gmail.com>"]
readme = "README.md"
packages = [{include = "digimon_card_game"}, {include = "tests"}]

[tool.poetry.dependencies]
python = "^3.11"
attrs = "^23.1.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
pytest-cov = "^4.1.0"
mypy = "1.5.1"
pre-commit = "3.3.3"

[[tool.mypy.overrides]]
module = ["digimon_card_game.*", "tests.*"]
implicit_reexport = true

[tool.ruff]
target-version = "py311"
select = ["ALL"]
ignore = [
    "COM812",   # missing-trailing-comma
    "D",        # pydocstyle
    "E501",     # line-too-long (we use black)
    "EM101",    # raw-string-in-exception
    "EM102",    # f-string-in-exception
    "ANN101",   # missing-type-self
    "TID252",   # relative-imports
    "TRY003",   # raise-vanilla-args (we do not need to create additional exceptions)
    "TCH",      # flake8-type-checking
]
force-exclude = true # to respect exclusions (useful for pre-commit)

[tool.ruff.per-file-ignores]
"__init__.py" = [ "F401" ] # unused-import
"tests/**/*.py" = [
    "S101",     # assert
    "PLR2004",  # magic-value-comparison
]

[tool.ruff.flake8-builtins]
builtins-ignorelist = ["type"] # digi-egg cards and digimon cards

[tool.ruff.mccabe]
max-complexity = 5

[tool.ruff.pep8-naming]
ignore-names = ["Date"] # we tend to rename the class

[tool.ruff.isort]
force-single-line = true # Forces all from imports to appear on their own line.

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
